Create a benchmarks.jar out of bootstrap jar for end user experience
* Having benchmarks.jar installed under end user application
provides with an experience just like with JMH.
* To achieve this, a new module fibula-benchmarks has been created
which only has the fibula-bootstrap project as dependency.
* This new module has an quarkus uber jar package type,
so when built it will take bootstrap and dependencies
and create an uber jar.
* Then, end user application would use the `maven-dependency-plugin`
to copy this uber jar dependency into the target folder as `benchmarks.jar`,
and that would be the starting point.