Avoid the need to override getForkedMainCommand
* Using a cutom output format enables
most of the command line arguments necessary for native to be passed in,
so the only thing that needed to be fixed is make sure not -XX: flags
are passed in because those are not understood in native.
* Relying on JMH's getForkedMainCommand implementation means
all the compiled generated JMH benchmarks need to be found in a directory
that is already covered by the classpath,
so decided to copy generated compiled bytecode into test-classes folder.